Add Open-Api UI + clean up

+++ src/main/java/com/thehrdivision/HelloRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HelloRepository extends JpaRepository<Hello, Long> {
+public interface HelloRepository extends JpaRepository<Hello, Integer> {
 }
